REF: Move `configuration`
